VACMS-21015: Moves new release listing to products
