+++ .github/CODEOWNERS
@@ -18,3 +18,4 @@ src/products/staffProfile/ @department-of-veterans-affairs/ap-team
 src/products/newsStory/ @department-of-veterans-affairs/ap-team
 src/products/storyListing/ @department-of-veterans-affairs/ap-team
 src/products/pressRelease/ @department-of-veterans-affairs/ap-team
+src/products/pressReleaseListing/ @department-of-veterans-affairs/ap-team
