+++ src/data/queries/index.ts
@@ -30,7 +30,7 @@ import * as NumberCallout from '@/data/queries/numberCallout'
 import * as PersonProfile from '@/products/staffProfile/query'
 import * as PhoneNumber from '@/data/queries/phoneNumber'
 import * as PressRelease from '@/products/pressRelease/query'
-import * as PressReleaseListing from '@/data/queries/pressReleaseListing'
+import * as PressReleaseListing from '@/products/pressReleaseListing/query'
 import * as PressReleaseTeaser from '@/data/queries/pressReleaseTeaser'
 import * as ProcessList from '@/data/queries/processList'
 import * as PromoBlock from '@/data/queries/promoBlock'
