+++ src/pages/[[...slug]].tsx
@@ -40,7 +40,7 @@ import { EventListing as FormattedEventListing } from '@/products/eventListing/f
 import { LocationsListing as FormattedLocationsListing } from '@/types/formatted/locationsListing'
 import { NewsStory as FormattedNewsStory } from '@/products/newsStory/formatted-type'
 import { PressRelease as FormattedPressRelease } from '@/products/pressRelease/formatted-type'
-import { PressReleaseListing as FormattedPressReleaseListing } from '@/types/formatted/pressReleaseListing'
+import { PressReleaseListing as FormattedPressReleaseListing } from '@/products/pressReleaseListing/formatted-type'
 import { ResourcesSupport as FormattedResourcesSupport } from '@/types/formatted/resourcesSupport'
 import { StaffProfile as FormattedStaffProfile } from '@/products/staffProfile/formatted-type'
 import { StoryListing as FormattedStoryListing } from '@/products/storyListing/formatted-type'
@@ -58,7 +58,7 @@ import { LocationsListing } from '@/templates/layouts/locationsListing'
 import { Meta } from '@/templates/common/meta'
 import { NewsStory } from '@/products/newsStory/template'
 import { PressRelease } from '@/products/pressRelease/template'
-import { PressReleaseListing } from '@/templates/layouts/pressReleaseListing'
+import { PressReleaseListing } from '@/products/pressReleaseListing/template'
 import { PreviewCrumb } from '@/templates/common/preview'
 import { ResourcesSupport } from '@/templates/layouts/resourcesSupport'
 import { StaffProfile } from '@/products/staffProfile/template'
