+++ src/products/pressReleaseListing/query.ts
@@ -1,9 +1,9 @@
 import { QueryData, QueryFormatter, QueryParams } from 'next-drupal-query'
-import { queries } from '.'
+import { queries } from '@/data/queries'
 import { DrupalJsonApiParams } from 'drupal-jsonapi-params'
 import { NodePressRelease, NodePressReleaseListing } from '@/types/drupal/node'
 import { Menu } from '@/types/drupal/menu'
-import { PressReleaseListing } from '@/types/formatted/pressReleaseListing'
+import { PressReleaseListing } from '@/products/pressReleaseListing/formatted-type'
 import { buildSideNavDataFromMenu } from '@/lib/drupal/facilitySideNav'
 import { ListingPageDataOpts } from '@/lib/drupal/listingPages'
 import { RESOURCE_TYPES } from '@/lib/constants/resourceTypes'
